flux/apps/games/futboss-ai/apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.10.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
mongomock==4.1.2

